Move common sql queries into separate files
